Mudei coisa pra caralho. Já nem sei o que fiz!

+++ process.c
@@ -3,10 +3,10 @@ struct Process{
     int ppid;
     int state;
     int priority;
-    float quantum;
+    int quantum;
 };
 
-struct Proccess *createProcess(int pid, int ppid, int priority, float quantum){
+struct Proccess *createProcess(int pid, int ppid, int priority, int quantum){
     struct Process *p = (struct Process*)malloc(sizeof(struct Process));
     p->pid = pid;
     p->ppid = ppid;
